Adjust for the future strictFunctionTypes, reduce Stream usage in gui
Alternative title: Lose all hope to adjust for strictFunctions

#4043

Turns out strictFunctions breaks not only streams but also pretty much
all generic components. Type inference doesn't work the way we expect
and doesn't fill in the class generic parameter, it infers
MyComponent<unknown> and then fails to unify it with
MyAttributes<Something>.

We don't care about attrs.state in pretty much all cases and it's
also unambigious what it is for class components so for us it's more
of a problem than a benefit.

We've added a generic helper which allows us to specify the type of
attrs manually without spelling out the whole class name again but it's
still not ideal that we have to specify even that. Alternatively we
could write a wrapper for each generic component e.g. `renderDropdown`
which would make type inference work but this is also a hack and is
unidiomatic.